Refactor ConcurrentWalSystem for readability
* Extracted BackgroundFlusher to src/storage/wal/background_flusher.rs
* Refactored ConcurrentWalSystem::new to use helper functions
* Improved code organization and readability without changing behavior
