+++ src/storage/wal/concurrent_system.rs
@@ -45,10 +45,10 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::{Condvar, Mutex};
 use std::thread::{self, JoinHandle};
 use std::time::Duration;
 
+use super::background_flusher::{BackgroundFlusher, FlushNotifier};
 use super::concurrent::{ConcurrentWal, ConcurrentWalConfig};
 use super::flush_coordinator::{FlushCoordinator, FlushCoordinatorConfig, FlushStats};
 use super::group_commit::GroupCommitCoordinator;
@@ -120,147 +120,6 @@ impl ConcurrentWalSystemConfig {
     }
 }
 
-/// Signal for waking up the flush thread when batch is full.
-struct FlushNotifier {
-    /// Lock for condvar.
-    lock: Mutex<bool>,
-    /// Condvar to signal immediate flush.
-    condvar: Condvar,
-}
-
-impl FlushNotifier {
-    fn new() -> Self {
-        Self {
-            lock: Mutex::new(false),
-            condvar: Condvar::new(),
-        }
-    }
-
-    /// Signal the flush thread to wake up immediately.
-    fn notify(&self) {
-        let mut guard = self.lock.lock().unwrap_or_else(|e| e.into_inner());
-        *guard = true;
-        self.condvar.notify_one();
-    }
-
-    /// Wait for a signal or timeout, returns true if signaled.
-    fn wait_timeout(&self, duration: Duration) -> bool {
-        let mut guard = self.lock.lock().unwrap_or_else(|e| e.into_inner());
-
-        // Check if already signaled before waiting.
-        // This handles the race where notify() is called before we enter wait_timeout().
-        if *guard {
-            *guard = false; // Reset signal
-            return true;
-        }
-
-        let (new_guard, result) = self
-            .condvar
-            .wait_timeout(guard, duration)
-            .unwrap_or_else(|e| e.into_inner());
-        guard = new_guard;
-
-        let was_signaled = *guard && !result.timed_out();
-        *guard = false; // Reset signal
-        was_signaled
-    }
-}
-
-/// Threshold for consecutive flush errors before logging a critical warning.
-const FLUSH_ERROR_WARNING_THRESHOLD: u64 = 3;
-
-/// Helper struct to encapsulate background flush logic.
-struct BackgroundFlusher {
-    wal: Arc<ConcurrentWal>,
-    coordinator: Arc<FlushCoordinator>,
-    shutdown: Arc<AtomicBool>,
-    flush_notifier: Arc<FlushNotifier>,
-    group_commit: Option<Arc<GroupCommitCoordinator>>,
-    error_counter: Arc<AtomicU64>,
-    interval: Duration,
-    sync_on_flush: bool,
-}
-
-impl BackgroundFlusher {
-    fn run(&self) {
-        while !self.shutdown.load(Ordering::Relaxed) {
-            self.perform_flush_cycle();
-            // Wait for flush interval OR immediate signal (batch full)
-            self.flush_notifier.wait_timeout(self.interval);
-        }
-        self.perform_final_flush();
-    }
-
-    fn perform_flush_cycle(&self) {
-        let entries = self.wal.drain_all();
-
-        // Always try to advance the epoch when there are entries OR when
-        // group commit has pending transactions.
-        //
-        // LOCK POISONING: If current_batch_size() fails, the coordinator lock is
-        // poisoned and the system is in an unrecoverable state. Panicking is correct
-        // here - continuing would leave waiting transactions hanging indefinitely.
-        let should_mark_flushed = !entries.is_empty()
-            || self.group_commit.as_ref().is_some_and(|gc| {
-                gc.current_batch_size()
-                    .expect("GroupCommitCoordinator lock poisoned - flush thread cannot continue")
-                    > 0
-            });
-
-        if !entries.is_empty() {
-            // Flush to coordinator
-            let result = self.coordinator.flush(entries, self.sync_on_flush);
-            self.handle_flush_result(result.map(|_| ()));
-        } else if should_mark_flushed {
-            // No entries but there are pending transactions - advance epoch anyway
-            self.handle_flush_result(Ok(()));
-        }
-    }
-
-    fn perform_final_flush(&self) {
-        let entries = self.wal.drain_all();
-        if !entries.is_empty() {
-            let result = self.coordinator.flush(entries, true);
-            self.handle_flush_result(result.map(|_| ()));
-        }
-    }
-
-    fn handle_flush_result(&self, result: Result<()>) {
-        match result {
-            Ok(_) => {
-                // Reset error counter on success
-                self.error_counter.store(0, Ordering::Relaxed);
-                if let Some(ref gc) = self.group_commit {
-                    gc.mark_flushed(Ok(())).expect(
-                        "GroupCommitCoordinator lock poisoned - flush thread cannot continue",
-                    );
-                }
-            }
-            Err(e) => {
-                // Track consecutive errors for health monitoring
-                let errors = self.error_counter.fetch_add(1, Ordering::Relaxed) + 1;
-                if errors == FLUSH_ERROR_WARNING_THRESHOLD {
-                    eprintln!(
-                        "CRITICAL: WAL flush failed {} consecutive times. \
-                         Data durability may be compromised. Last error: {}",
-                        errors, e
-                    );
-                } else {
-                    eprintln!("WAL flush error: {}", e);
-                }
-
-                if let Some(ref gc) = self.group_commit {
-                    // Create a new error from the string representation
-                    gc.mark_flushed(Err(crate::core::error::Error::other(e.to_string())))
-                        .expect(
-                            "GroupCommitCoordinator lock poisoned - flush thread cannot continue",
-                        );
-                }
-            }
-        }
-    }
-}
-
 /// Unified concurrent WAL system.
 ///
 /// This combines the striped concurrent WAL with the flush coordinator
@@ -287,7 +146,42 @@ pub struct ConcurrentWalSystem {
 impl ConcurrentWalSystem {
     /// Create a new concurrent WAL system.
     pub fn new(config: ConcurrentWalSystemConfig) -> Result<Self> {
-        // Create ConcurrentWal config
+        let (wal_config, coordinator_config) = Self::create_configs(&config);
+
+        let wal = Arc::new(ConcurrentWal::new(wal_config)?);
+        let coordinator = Arc::new(FlushCoordinator::new(coordinator_config)?);
+        let shutdown_signal = Arc::new(AtomicBool::new(false));
+
+        let group_commit = Self::create_group_commit(config.durability_mode);
+        let flush_notifier = Arc::new(FlushNotifier::new());
+        let consecutive_flush_errors = Arc::new(AtomicU64::new(0));
+
+        let flush_thread = Self::spawn_flush_thread(
+            config.durability_mode,
+            config.flush_interval_ms,
+            Arc::clone(&wal),
+            Arc::clone(&coordinator),
+            Arc::clone(&shutdown_signal),
+            Arc::clone(&flush_notifier),
+            group_commit.clone(),
+            Arc::clone(&consecutive_flush_errors),
+        );
+
+        Ok(Self {
+            wal,
+            coordinator,
+            flush_thread,
+            shutdown_signal,
+            flush_notifier,
+            durability_mode: config.durability_mode,
+            group_commit,
+            consecutive_flush_errors,
+        })
+    }
+
+    fn create_configs(
+        config: &ConcurrentWalSystemConfig,
+    ) -> (ConcurrentWalConfig, FlushCoordinatorConfig) {
         let wal_config = ConcurrentWalConfig {
             wal_dir: config.wal_dir.clone(),
             num_stripes: config.num_stripes,
@@ -296,9 +190,8 @@ impl ConcurrentWalSystem {
             segments_to_retain: config.segments_to_retain,
         };
 
-        // Create FlushCoordinator config
         let coordinator_config = FlushCoordinatorConfig {
-            wal_dir: config.wal_dir,
+            wal_dir: config.wal_dir.clone(),
             segment_size: config.segment_size,
             segments_to_retain: config.segments_to_retain,
             flush_interval_ms: config.flush_interval_ms,
@@ -308,13 +201,11 @@ impl ConcurrentWalSystem {
             ),
             write_buffer_size: config.write_buffer_size,
         };
+        (wal_config, coordinator_config)
+    }
 
-        let wal = Arc::new(ConcurrentWal::new(wal_config)?);
-        let coordinator = Arc::new(FlushCoordinator::new(coordinator_config)?);
-        let shutdown_signal = Arc::new(AtomicBool::new(false));
-
-        // Create group commit coordinator for modes that need epoch tracking
-        let group_commit = match config.durability_mode {
+    fn create_group_commit(mode: DurabilityMode) -> Option<Arc<GroupCommitCoordinator>> {
+        match mode {
             DurabilityMode::GroupCommit {
                 max_batch_size,
                 max_delay_ms,
@@ -331,57 +222,44 @@ impl ConcurrentWalSystem {
                 max_batch_size,
             ))),
             _ => None,
-        };
+        }
+    }
 
-        // Create flush notifier for batch-size-triggered flushes
-        let flush_notifier = Arc::new(FlushNotifier::new());
-
-        // Create error counter for health monitoring
-        let consecutive_flush_errors = Arc::new(AtomicU64::new(0));
-
-        // Start background flush thread for async/group-commit modes
-        let flush_thread = if matches!(
-            config.durability_mode,
+    #[allow(clippy::too_many_arguments)]
+    fn spawn_flush_thread(
+        durability_mode: DurabilityMode,
+        flush_interval_ms: u64,
+        wal: Arc<ConcurrentWal>,
+        coordinator: Arc<FlushCoordinator>,
+        shutdown: Arc<AtomicBool>,
+        flush_notifier: Arc<FlushNotifier>,
+        group_commit: Option<Arc<GroupCommitCoordinator>>,
+        error_counter: Arc<AtomicU64>,
+    ) -> Option<JoinHandle<()>> {
+        if matches!(
+            durability_mode,
             DurabilityMode::Async { .. }
                 | DurabilityMode::GroupCommit { .. }
                 | DurabilityMode::AsyncBatched { .. }
         ) {
-            let wal_clone = Arc::clone(&wal);
-            let coordinator_clone = Arc::clone(&coordinator);
-            let shutdown_clone = Arc::clone(&shutdown_signal);
-            let flush_notifier_clone = Arc::clone(&flush_notifier);
-            let group_commit_clone = group_commit.clone();
-            let error_counter_clone = Arc::clone(&consecutive_flush_errors);
-            let flush_interval = Duration::from_millis(config.flush_interval_ms);
-            let sync_on_flush =
-                matches!(config.durability_mode, DurabilityMode::GroupCommit { .. });
+            let flush_interval = Duration::from_millis(flush_interval_ms);
+            let sync_on_flush = matches!(durability_mode, DurabilityMode::GroupCommit { .. });
 
             Some(thread::spawn(move || {
                 Self::flush_loop(
-                    wal_clone,
-                    coordinator_clone,
-                    shutdown_clone,
-                    flush_notifier_clone,
-                    group_commit_clone,
-                    error_counter_clone,
+                    wal,
+                    coordinator,
+                    shutdown,
+                    flush_notifier,
+                    group_commit,
+                    error_counter,
                     flush_interval,
                     sync_on_flush,
                 );
             }))
         } else {
             None
-        };
-
-        Ok(Self {
-            wal,
-            coordinator,
-            flush_thread,
-            shutdown_signal,
-            flush_notifier,
-            durability_mode: config.durability_mode,
-            group_commit,
-            consecutive_flush_errors,
-        })
+        }
     }
 
     /// Background flush loop.
